# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Inventory API",
        "description": "REST API for managing coffee shop inventory, products and stock levels.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://brew-control.test",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products (Read)"
                ],
                "summary": "List all products",
                "description": "Returns a list of all products available in the inventory",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products (Write)"
                ],
                "summary": "Update a product",
                "description": "Updates an existing product in the inventory. **Note:** You must include the `id` of the product you want to update.",
                "operationId": "86ea040208eec2e4c24ce8a37f8fffc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products (Write)"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product in the inventory",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search/name/{name}": {
            "get": {
                "tags": [
                    "Products (Read)"
                ],
                "summary": "Get a product by name",
                "description": "Returns a product that matches the given name",
                "operationId": "c2d142ecf2a359fe6e310e6e112e4fb6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the product to search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search/category/{category}": {
            "get": {
                "tags": [
                    "Products (Read)"
                ],
                "summary": "Get products by category",
                "description": "Returns all products belonging to a specific category",
                "operationId": "620a3d143dd257b69ccb45463e60ec33",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category name to filter products",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{name}": {
            "delete": {
                "tags": [
                    "Products (Write)"
                ],
                "summary": "Delete a product by name",
                "description": "Deletes the product with the specified name from the inventory",
                "operationId": "f3a87e6f27fdac8d8120c8163fb31a4e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductRequest": {
                "description": "Data required to create or update a product",
                "required": [
                    "name",
                    "category",
                    "price",
                    "quantity"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the product",
                        "type": "string",
                        "example": "Espresso Coffee"
                    },
                    "category": {
                        "description": "The category of the product (e.g., Beans, Capsules)",
                        "type": "string",
                        "example": "Beans"
                    },
                    "price": {
                        "description": "The price of the product in local currency",
                        "type": "number",
                        "format": "float",
                        "example": 15.9
                    },
                    "quantity": {
                        "description": "Available quantity in stock",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ProductResponse": {
                "description": "Representation of a product in the inventory",
                "properties": {
                    "id": {
                        "description": "Unique product ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Espresso Coffee"
                    },
                    "category": {
                        "description": "Product category",
                        "type": "string",
                        "example": "Beans"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": 15.9
                    },
                    "quantity": {
                        "description": "Stock quantity",
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "description": "Record creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Record last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products (Read)",
            "description": "Products (Read)"
        },
        {
            "name": "Products (Write)",
            "description": "Products (Write)"
        }
    ]
}